fixed norme errors
