+++ src/error_handling/errors2.c
@@ -1,26 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   frees2.c                                           :+:      :+:    :+:   */
+/*   errors2.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/28 17:51:12 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/30 18:37:41 by mortins-         ###   ########.fr       */
+/*   Created: 2023/10/30 18:54:52 by mortins-          #+#    #+#             */
+/*   Updated: 2023/10/30 18:58:00 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/minishell.h"
+#include "../../inc/minishell.h"
 
-void	free_list_and_data(t_list **list)
+int	export_error_msg(char *error)
 {
-	t_list	*tmp;
-
-	while (*list)
-	{
-		tmp = *list;
-		*list = (*list)->next;
-		free(tmp->data);
-		free(tmp);
-	}
+	g_exit = 1;
+	ft_putstr_fd("Minishell: export: '", STDERR_FILENO);
+	ft_putstr_fd(error, STDERR_FILENO);
+	ft_putstr_fd("': not a valid identifier\n", STDERR_FILENO);
+	return (1);
 }
