+++ src/run/exec.c
@@ -6,7 +6,7 @@
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 17:13:44 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/30 18:25:39 by mortins-         ###   ########.fr       */
+/*   Updated: 2023/10/30 19:10:06 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,54 +37,6 @@ void	exec(t_minishell *ms, char **cmd_arr)
 	free(cmd_path);
 	g_exit = errno;
 	free_ms(ms);
-}
-
-int	is_usable(char	*cmd, char *cmd_path, char **paths_array)
-{
-	struct stat path_stat;
-
-	stat(cmd_path, &path_stat);
-	if (S_ISDIR(path_stat.st_mode) != 0)
-	{
-		ft_putstr_fd("Minishell: ", STDERR_FILENO);
-		ft_putstr_fd(cmd, STDERR_FILENO);
-		ft_putstr_fd(": Is a directory", STDERR_FILENO);
-		g_exit = 126;
-	}
-	else if (access(cmd_path, X_OK) == 0)
-	{
-		free(cmd_path);
-		return (1);
-	}
-	else  if (access(cmd_path, X_OK) != 0)
-	{
-		perror("Minishell: error: ");
-		g_exit = 126;
-	}
-	else
-		g_exit = 1;
-	free(cmd_path);
-	free_array(paths_array);
-	return (0);
-}
-
-int	is_exec(char *cmd, char **paths)
-{
-	int		i;
-	char	*buf;
-	char	*buf2;
-
-	i = 0;
-	while (paths[i])
-	{
-		buf = ft_strjoin(paths[i++], "/");
-		buf2 = ft_strjoin(buf, cmd);
-		free(buf);
-		if (access(buf2, F_OK) == 0)
-			return (is_usable(cmd, buf2, paths));
-		free(buf2);
-	}
-	return (1);
 }
 
 char	**special_path(const char *cmd)
@@ -119,7 +71,7 @@ char	**get_paths(t_list **env, char *cmd)
 	char	**path_dir;
 	int		i;
 
-	if (ft_strncmp(cmd, "../", 3) == 0 || ft_strncmp(cmd, "./", 2 ) == 0 \
+	if (ft_strncmp(cmd, "../", 3) == 0 || ft_strncmp(cmd, "./", 2) == 0 \
 		|| cmd[0] == '/')
 		return (special_path(cmd));
 	tmp = *env;
