+++ src/run/run.c
@@ -6,7 +6,7 @@
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/23 16:01:57 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/30 18:33:00 by mortins-         ###   ########.fr       */
+/*   Updated: 2023/10/30 18:44:40 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -106,11 +106,8 @@ void	parent(t_minishell *ms, int *pipe_fd, int cmds_run, int pos)
 
 	cmd = ms->cmdlist;
 	i = cmds_run;
-	while (i > 0)
-	{
+	while (i-- > 0)
 		cmd = cmd->next;
-		i--;
-	}
 	if (ms->cmd_count == 1)
 	{
 		if (is_built_in(cmd->content->cmd_flags[0]))
