+++ src/signals.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   signals.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ddiniz-m <ddiniz-m@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/16 16:08:59 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/30 14:59:42 by ddiniz-m         ###   ########.fr       */
+/*   Updated: 2023/10/30 18:38:23 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,10 +42,7 @@ void	signal_process_interrupt(int signum)
 void	signal_exit(t_minishell *ms)
 {
 	if (!(ms->str))
-	{
-		/* write(2, "exit\n", 5); */
 		free_ms(ms);
-	}
 }
 
 //	CTRL-C sends an "Interrupt" signal.
