+++ src/structs/content.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   content.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ddiniz-m <ddiniz-m@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/14 14:59:23 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/30 14:45:17 by ddiniz-m         ###   ########.fr       */
+/*   Updated: 2023/10/30 18:47:42 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,11 +78,11 @@ char	**cmd_with_flags(t_minishell *ms, char **arr, int pos)
 	i = 0;
 	args = cmd_args(arr, pos);
 	buf = malloc(sizeof(char *) * (args + 1));
-	if (!buf)	
+	if (!buf)
 		malloc_error(ms);
 	while (i < args && arr[pos])
 	{
-		if (ft_strcmp(arr[pos], ">") == 0 || ft_strcmp(arr[pos], ">>") == 0\
+		if (ft_strcmp(arr[pos], ">") == 0 || ft_strcmp(arr[pos], ">>") == 0 \
 			|| ft_strcmp(arr[pos], "<") == 0 || ft_strcmp(arr[pos], "<<") == 0)
 			pos += 2;
 		else if (ft_strcmp(arr[pos], "|") == 0)
